fix: user applyEdits to correctly create intermediate latex
