+++ src/parser/types.ts
@@ -1,3 +1,5 @@
+import type { Replacement } from "./utility";
+
 export type FieldTypes = "math" | "plot" | "parameter" | "units" | "expression" | "number" |
   "condition" | "piecewise" | "expression_no_blank" | "equality" | "id_list" | "data_table_expression" |
   "data_table_assign";
@@ -190,7 +192,7 @@ export type QueryStatement = BaseQueryStatement & {
   isCodeFunctionQuery: false;
   isCodeFunctionRawQuery: false;
   subQueries: SubQueryStatement[];
-  subQueryToLatexMap: Map<string,string>;
+  subQueryReplacements: [string, Replacement][];
   isSubQuery: false;
 };
 
@@ -204,13 +206,13 @@ export type DataTableQueryStatement = BaseQueryStatement & {
   colNum: number;
 };
 
-export type EqualityUnitsQueryStatement = Omit<QueryStatement, "unitsLatex" | "dimensions" | "subQueries" | "subQueryToLatexMap" | "isSubQuery"> & {
+export type EqualityUnitsQueryStatement = Omit<QueryStatement, "unitsLatex" | "dimensions" | "subQueries" | "subQueryReplacements" | "isSubQuery"> & {
   isEqualityUnitsQuery: true;
   isSubQuery: false;
   equationIndex: number;
 };
 
-export type SubQueryStatement = Omit<QueryStatement, "isSubQuery" | "subQueries" | "subQueryToLatexMap" > & {
+export type SubQueryStatement = Omit<QueryStatement, "isSubQuery" | "subQueries" | "subQueryReplacements" > & {
   isSubQuery: true;
 };
 
@@ -241,13 +243,13 @@ export type ParametricRangeQueryStatement = {
   rangeQueryStatements: RangeQueryStatement[];
 }
 
-export type ScatterXValuesQueryStatement = Omit<QueryStatement, "isScatterXValuesQueryStatement" | "subQueries" | "subQueryToLatexMap" | "isSubQuery"> & {
+export type ScatterXValuesQueryStatement = Omit<QueryStatement, "isScatterXValuesQueryStatement" | "subQueries" | "subQueryReplacements" | "isSubQuery"> & {
   isScatterXValuesQueryStatement: true;
   isSubQuery: false;
   equationIndex: number;
 }
 
-export type ScatterYValuesQueryStatement = Omit<QueryStatement, "isScatterYValuesQueryStatement" | "subQueries" | "subQueryToLatexMap" | "isSubQuery"> & {
+export type ScatterYValuesQueryStatement = Omit<QueryStatement, "isScatterYValuesQueryStatement" | "subQueries" | "subQueryReplacements" | "isSubQuery"> & {
   isScatterYValuesQueryStatement: true;
   isSubQuery: false;
   equationIndex: number;
@@ -318,27 +320,6 @@ export type FunctionUnitsQuery = Pick<BaseQueryStatement, "type" | "sympy" | "pa
   isCodeFunctionQuery: false;
   isCodeFunctionRawQuery: false;
 };
-
-export type Insertion = {
-  type: "insertion";
-  location: number;
-  text: string;
-};
-
-export function isInsertion(edit: (Insertion | Replacement)): edit is Insertion {
-  return edit.type === "insertion";
-}
-
-export type Replacement = {
-  type: "replacement";
-  location: number;
-  deletionLength: number;
-  text: string
-}
-
-export function isReplacement(edit: (Insertion | Replacement)): edit is Replacement {
-  return edit.type === "replacement";
-}
 
 export type DataTableInfo = {
   colVars: string[];
